requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
behave==1.2.6
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cheroot==10.0.1
ci-info==0.3.0
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
config==0.5.1
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
cssselect==1.2.0
curl_cffi==0.6.4
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.2.0
entrypoints==0.4
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.2
execnet==2.1.1
fake-useragent==1.5.1
Faker==26.0.0
fastapi==0.111.0
fastapi-cli==0.0.4
fasteners==0.19
favicon==0.7.0
ffmpeg-python==0.2.0
filelock==3.15.4
fitz==0.0.1.dev2
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gcsfs==2024.6.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.49.0
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-cloud-vision==3.7.2
google-crc32c==1.5.0
google-generativeai==0.6.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isodate==0.6.1
jaraco.functools==4.0.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
looseversion==1.3.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
more-itertools==10.3.0
MouseInfo==0.1.3
multidict==6.0.5
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
oauthlib==3.2.2
openai==1.33.0
openpyxl==3.1.3
orjson==3.10.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
parse==1.20.2
parse-type==0.6.2
pathlib==1.0.1
pdbp==1.5.3
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
proto-plus==1.23.0
protobuf==4.25.3
prov==2.0.1
py==1.11.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
pycurl==7.45.3
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
pydot==2.0.0
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.18.0
pymdown-extensions==10.8.1
pymongo==4.8.0
PyMsgBox==1.0.9
pynose==1.5.2
# pyobjc-core==10.3.1
# pyobjc-framework-Cocoa==10.3.1
# pyobjc-framework-Quartz==10.3.1
pyotp==2.9.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytest==8.3.2
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytweening==1.2.0
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==6.3.2
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
rubicon-objc==0.4.9
sbvirtualdisplay==1.3.0
scipy==1.14.0
selenium==4.23.1
selenium-base==4.29.3
Selenium-Screenshot==2.1.0
selenium-stealth==1.0.6
seleniumbase==4.29.3
shapely==2.0.4
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.4
st-annotated-text==4.0.1
st-theme==1.2.3
starlette==0.37.2
streamlit==1.36.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tabcompleter==1.3.3
tenacity==8.4.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traits==6.3.2
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
ujson==5.10.0
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.2.2
user-agents==2.2.0
uvicorn==0.30.1
uvloop==0.19.0
validators==0.29.0
vertexai==1.49.0
watchfiles==0.22.0
web.py==0.62
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==12.0
wget==3.2
wsproto==1.2.0
yarl==1.9.4
